# doc/api.yaml
openapi: 3.0.0
servers:
  - url: http://localhost:3000/api/v3
info:
  title: WASAText
  description: Web And Software Architecture Exam Project
  version: 1.0.0
security:
  - bearerAuth: []
...
paths:
...
  /groups/{groupId}/message:
    parameters:
      - name: groupId
        in: path
        description: Group id
        required: true
        schema:
          type: integer
          format: int64
    post:
      tags: ["groups"]
      summary: Send a message to a group
      description: Send a message to a group
      operationId: sendMessageToGroup
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/MessageContent"
      responses:
        '201':
          description: Created
          content:
            application/json:
              example:
                type: object
              schema:
                $ref: "#/components/schemas/Message"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
  /conversations/{conversationId}:
    parameters:
      - name: conversationId
        in: path
        description: Conversation id
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: ["conversations"]
      summary: Get conversation by id
      description: Get conversation by id
      operationId: getConversation
      responses:
        '200':
          description: Successful
          content:
            application/json:
              example:
                type: object
              schema:
                description: Conversation messages
                type: object
                properties:
                  messageIds:
                    description: Message ids
                    type: array
                    minItems: 1
                    maxItems: 1024
                    items:
                      description: Message id
                      type: integer
                      format: int64
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
  /reactions/{messageId}:
    parameters:
      - name: messageId
        in: path
        description: Message id
        required: true
        schema:
          type: integer
          format: int64
    post:
      tags: ["reactions"]
      summary: Add reaction to a message
      description: Add reaction to a message
      operationId: addReaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: Emoji
              type: object
              properties:
                emoji:
                  $ref: "#/components/schemas/Emoji"
      responses:
        '201':
          description: Created
          content:
            application/json:
              example:
                type: object
              schema:
                $ref: "#/components/schemas/Reaction"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: ["reactions"]
      summary: Remove a reaction from a message
      description: Remove a reaction from a message
      operationId: deleteReaction
      responses:
        '204':
          description: No Content
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
    get:
      tags: ["reactions"]
      summary: Get message reactions
      description: Get message reactions
      operationId: getReactions
      responses:
        '200':
          description: Successful
          content:
            application/json:
              example:
                type: object
              schema:
                description: Message reactions
                type: object
                properties:
                  reactions:
                    description: Reaction ids
                    type: array
                    minItems: 1
                    maxItems: 1024
                    items:
                      $ref: "#/components/schemas/Reaction"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /comments/{messageId}:
    parameters:
      - name: messageId
        in: path
        description: Message id
        required: true
        schema:
          type: integer
          format: int64
    delete:
      tags: ["comments"]
      summary: Remove a comment
      description: Remove a comment
      operationId: uncommentMessage
      responses:
        '204':
          description: No Content
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
    post:
      tags: ["comments"]
      summary: Create a comment
      description: Create a comment
      operationId: commentMessage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/MessageContent"
      responses:
        '201':
          description: Created
          content:
            application/json:
              example:
                type: object
              schema:
                $ref: "#/components/schemas/Message"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
  /messages/{messageId}:
    parameters:
      - name: messageId
        in: path
        description: Message id
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: ["messages"]
      operationId: getMessage
      summary: Get message data
      description: Get message data
      responses:
        '200':
          description: Successful
          content:
            application/json:
              example:
                type: object
              schema:
                $ref: "#/components/schemas/Message"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: ["messages"]
      operationId: deleteMessage
      summary: Delete message
      description: Delete message
      responses:
        '204':
          description: No Content
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
  /users/{userId}/message:
    parameters:
      - name: userId
        in: path
        description: User id
        required: true
        schema:
          type: integer
          format: int64
    post:
      tags: ["users"]
      summary: Send a message to a user
      description: Send a message to a user
      operationId: sendMessage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/MessageContent"
      responses:
        '201':
          description: Created
          content:
            application/json:
              example:
                type: object
              schema:
                $ref: "#/components/schemas/Message"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}/fmessage:
    parameters:
      - name: userId
        in: path
        description: User id
        required: true
        schema:
          type: integer
          format: int64
    post:
      tags: ["users"]
      summary: Forward a message to a user
      description: Forward a message to a group
      operationId: forwardMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: Request body
              type: object
              properties:
                messageId:
                  description: The message to forward
                  type: integer
                  format: int64
      responses:
        '201':
          description: Created
          content:
            application/json:
              example:
                type: object
              schema:
                $ref: "#/components/schemas/Message"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            description: Error
            type: object
            properties:
              error:
                description: Message
                type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            description: Error
            type: object
            properties:
              error:
                description: Message
                type: string
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            description: Error
            type: object
            properties:
              error:
                description: Message
                type: string
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            description: Error
            type: object
            properties:
              error:
                description: Message
                type: string
  schemas:
...
    Message:
      description: Message
      type: object
      properties:
        messageId:
          description: Message id
          type: integer
          format: int64
        text:
          description: Message text
          type: string
          pattern: '^.*$'
          minLength: 1
          maxLength: 1024
        senderId:
          description: Sender id
          type: integer
          format: int64
        conversationId:
          description: Conversation id
          type: integer
          format: int64
        createdAt:
          description: Message timestamp
          type: string
          format: date-time
        isForwarded:
          description: Message is forwarded
          type: boolean
        commentTo:
          description: Commented message id
          type: integer
          format: int64
        attachmentUrl:
          description: Attachment URL
          type: string
          format: url
        mediaType:
          $ref: "#/components/schemas/MediaType"
    MessageContent:
      description: Message content
      type: object
      properties:
        text:
          description: Message text
          type: string
          minLength: 0
          maxLength: 1024
          pattern: "^.*$"
        file:
          description: Binary file
          type: string
          format: binary
          minLength: 1
          maxLength: 1024
        mediaType:
          $ref: "#/components/schemas/MediaType"
    MediaType:
      description: Media type
      type: string
      enum: [image, video, audio, file]
...
    Reaction:
      description: Reaction
      type: object
      properties:
        emoji:
          $ref: "#/components/schemas/Emoji"
        messageId:
          description: Message id
          type: integer
          format: int64
        senderId:
          description: Sender id
          type: integer
          format: int64
...
    Emoji:
      description: Emoji
      type: string
      enum: [like, love, laugh, sad, angry, thumb]